# perf(portex): use "getattr" to replace "hasattr" in "PortexType.imports"