| cases           | before           | after            |
| --------------- | ---------------- | ---------------- |
| bdd100k.imports | 787 µs ± 9.05 µs | 249 µs ± 4.39 µs |