@GravitiRob 2022-05-28
Task linked: [CU-2e11dkx 使用 getattr 代替 hasattr 提高 PortexType.imports 的效率](https://app.clickup.com/t/2e11dkx)